rename ExapnsionToggleDynamicIcon to ExpansionToggleDynamicIcon
