+++ resources/js/components/Icons/ExpansionToggleDynamicIcon.vue
